pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langchain-qdrant>=0.2.1",
    "pydantic>=2.12.0",
    "pymongo>=4.15.3",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "rq==1.14.0",
    "streamlit>=1.50.0",
]
